refactor: controller, domain, view 형태로 패키지 분리
