--- src/main/kotlin/minesweeper/Ground.kt
@@ -1,11 +1,0 @@
-package minesweeper
-
-class Ground(height: Int, vertical: Int) {
-    init {
-        require(height > MINIMUM_SIZE && vertical > MINIMUM_SIZE)
-    }
-
-    companion object {
-        const val MINIMUM_SIZE = 0
-    }
-}
