+++ src/main/kotlin/minesweeper/domain/PositionGenerator.kt
@@ -1,5 +1,6 @@
-package minesweeper
+package minesweeper.domain
 
 interface PositionGenerator {
+
     fun generate(end: Int): Int
 }
